TCU: added per-tile memory-bandwidth limit.
We now have a per-tile external register that controls the max. usable
memory bandwidth of the tile. To enforce that, we delay NoC requests
correspondingly to cover both cache misses and TCU transfers.

The memory-bandwidth register consists of a rate, a limit, and an
amount. The rate defines how fast the amount is refilled in bytes per
millisecond. The TCU uses a periodic event for refilling. The limit
defines the maximum value of amount, whereas amount defines the
currently available bytes that can be transferred without throttling.

If the number of bytes to transfer is less than the amount, we simply
reduce amount by that number and perform the transfer immediately. If
the amount does not suffice, we use an overflow register to pay for the
transfer later and delay the transfer correspondingly. During refill,
the amount is not changed until the overflow has been reduced to 0.